refactor(test): move test utils from src to test directory
