+++ src/components/BoardControls/Export.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, screen } from '@testing-library/react';
 
-import { renderWithProviders, updateStore } from '../../utils/test';
+import { renderWithProviders, updateStore } from '../../../test/utils';
 import Export from './Export';
 
 const { clipboard } = navigator;
