+++ src/components/Columns/hooks/useLikes.test.tsx
@@ -1,13 +1,17 @@
 import { onValue } from 'firebase/database';
 
 import {
+  renderWithProviders,
+  store,
+  updateStore,
+} from '../../../../test/utils';
+import {
   BOARD_TEST_ID as boardId,
   ITEM_TEST_ID as itemId,
   USER_TEST_ID as userId,
 } from '../../../constants/test';
 import { getLikesRef } from '../../../firebase';
 import { Likes } from '../../../types';
-import { renderWithProviders, store, updateStore } from '../../../utils/test';
 import { useLikes } from './useLikes';
 
 const mockedOnValue = jest.mocked(onValue);
