+++ src/components/Item/Item.test.tsx
@@ -1,11 +1,11 @@
 import { fireEvent, screen } from '@testing-library/react';
 
+import { renderWithProviders, store, updateStore } from '../../../test/utils';
 import {
   BOARD_TEST_ID as boardId,
   COLUMN_TEST_ID as columnId,
   ITEM_TEST_ID as itemId,
 } from '../../constants/test';
-import { renderWithProviders, store, updateStore } from '../../utils/test';
 import Item from './Item';
 
 const props = {
