+++ src/components/Items/Items.test.tsx
@@ -1,10 +1,10 @@
 import { fireEvent, screen } from '@testing-library/react';
 
+import { renderWithProviders, updateStore } from '../../../test/utils';
 import {
   BOARD_TEST_ID as boardId,
   COLUMN_TEST_ID as columnId,
 } from '../../constants/test';
-import { renderWithProviders, updateStore } from '../../utils/test';
 import Items from './Items';
 
 describe('add item', () => {
