+++ src/components/Layout/Layout.test.tsx
@@ -1,4 +1,4 @@
-import { renderWithProviders } from '../../utils/test';
+import { renderWithProviders } from '../../../test/utils';
 import Layout from './Layout';
 
 it('renders without crashing', () => {
