+++ src/hooks/useAuth.test.tsx
@@ -1,12 +1,12 @@
 import { renderHook } from '@testing-library/react';
 import type { Unsubscribe, User } from 'firebase/auth';
 
+import { store, wrapper } from '../../test/utils';
 import {
   USER_TEST_EMAIL as userEmail,
   USER_TEST_ID as userId,
 } from '../constants/test';
 import { logEvent, onAuthStateChanged, signInAnonymously } from '../firebase';
-import { store, wrapper } from '../utils/test';
 import { useAuth } from './useAuth';
 
 jest.mock('../firebase', () => ({
