+++ src/pages/Board/Board.test.tsx
@@ -3,13 +3,13 @@ import type { DatabaseReference, DataSnapshot } from 'firebase/database';
 import { onValue } from 'firebase/database';
 import { useParams } from 'react-router-dom';
 
+import { renderWithProviders, router, updateStore } from '../../../test/utils';
 import {
   BOARD_TEST_ID as boardId,
   USER_TEST_ID as userId,
 } from '../../constants/test';
 import { getBoardDataRef } from '../../firebase';
 import { Board as BoardType } from '../../types';
-import { renderWithProviders, router, updateStore } from '../../utils/test';
 import Board from './Board';
 
 const mockedOnValue = jest.mocked(onValue);
