+++ src/pages/Boards/BoardCard.test.tsx
@@ -1,10 +1,10 @@
 import { fireEvent, screen } from '@testing-library/react';
 
+import { renderWithProviders, store, updateStore } from '../../../test/utils';
 import {
   BOARD_TEST_ID as boardId,
   DATE_NOW as dateNow,
 } from '../../constants/test';
-import { renderWithProviders, store, updateStore } from '../../utils/test';
 import BoardCard from './BoardCard';
 
 it('renders "Open board" link', () => {
