+++ src/pages/Logout/Logout.test.tsx
@@ -1,12 +1,12 @@
 import { waitFor } from '@testing-library/react';
 
-import { logEvent, signOut } from '../../firebase';
 import {
   renderWithProviders,
   router,
   store,
   updateStore,
-} from '../../utils/test';
+} from '../../../test/utils';
+import { logEvent, signOut } from '../../firebase';
 import Logout from './Logout';
 
 jest.mock('../../firebase', () => ({
